Added the functionality of search maquina by an identifier in corte_registro.php
Implemented a new input in corte_registro.php to introduce the identifier to search for the machine with that.
Then solved the problem with filling the inputs, it now needs to have every input filled, but when you are
making the registration, it takes the values from a $_SESSION['cortesValues'] and not from the inputs. The inputs
are readen just at the begining when validating the information.

Improved the information that display the cards in corte.php, making it more understandable.