Integrated changes done in administrador into trabajador
Pasted the file of maquina.php into base folder for trabajador.

Modified the headers in template, changed the links to redirect to the new
file maquina.php outside admin folder.

Pasted the logic of corte_registro in admin to trabajador.

Pasted the logic of maquina_registro in admin to trabajador.

The view of maquina is only a readonly.