Updated methods from conexion.php and corte.php
1) Updated the method of generateDateRangeQuery().
It can use inner joins, to use them, you have to prepare the
parameters where you are calling it, but remember that
before it, there is a boundary that makes the call.

For example:

$params = [
    'table' =>[
        ['name' => 'corte'],
        ['name' => 'maquina',  ---> here you make the joins
            'on' =>['corte.idMaq', 'maquina.idMaq']]
    ],
    'fechaFin' => [
        'dia'=>$diaFinal,
        'mes'=>$mesFinal,
        'anio'=>$anio
    ]
];

And then just call the method searchByDate()

2) Changed the logic from corte, it used to call
the methods directly from conection, now it uses the
boundary class.

3) Implemented the logic to search by and display the
identificador information into the cards